refactor: centralize util function and types
